Got invertible invariants working
Can now generate a set of invariants (of degree 1, 2 and 3) that can be
inverted in closed from (modulo global orientation) using an algorithm
from Joel Kileel.
Thus far the invariants are not linearly (nor algebraically) independent
and, therefore, there are some redundancies.  These will be removed as
time goes on.  Initially we will start with the linear dependencies and
then hopefully move on to the polynomial dependencies.