Made progress on inverting invariants
We now have basic support for inverting invariants of systems with
symmetry.  The basic problem is that the Gram matrix from I2 invariants
will be rank deficient when there are symmetries.  The way to deal with
this is to march up to higher l Gram matrices before switching to I3.
This invariably requires more radial basis functions as the minimum
required is 2l + 1.