Fixed mistake in descriptor with cutoff function
The cutoff function was using the 'real space' cutoff value, however,
prior to the cutoff function being called the environment was _already_
being scaled to fit within the unit sphere.  This led to a mismatch
where the cutoff function was working on the wrong scale.  This has been
fixed and performance of the descriptor is greatly improved as a result.